Stepped back a bit to rethink -- basic install won't do much now other than setting up pam and NSS.  With some manual configuration this is still a pretty OK setup script for server functionality.

+++ remove.sh
@@ -8,41 +8,44 @@ fi
 
 echo "Killing all active php5 instances..."
 if [ "`ps -A | grep php5-cgi | head -1`" ]; then
-	killall -9 php5-cgi
+	service php5-fpm stop
 fi
 
 echo "Removing NGINX config and restarting..."
 rm /etc/nginx/sites-enabled/inkspot
 rm /etc/nginx/sites-available/inkspot
 if [ "`ps -A | grep nginx | head -1`" ]; then
-	/etc/init.d/nginx restart
+	service nginx stop
 fi
-
-echo "Removing PowerDNS config and restarting..."
-rm /etc/powerdns/pdns.d/inkspot.conf
-/etc/init.d/pdns restart
 
 echo "Removing pgsql from nsswitch.conf..."
 rpl -q " pgsql" "" /etc/nsswitch.conf
 
-echo "Remvoing PAM module..."
+echo "Removing PAM module..."
 pam-auth-update --package --remove pgsql
 rm /etc/pam_pgsql.conf
 
-echo "Removing database and database users..."
-echo "DROP DATABASE inkspot" | sudo -u postgres psql
-echo "DROP USER inkspot"     | sudo -u postgres psql
-echo "DROP USER inkspot_ro"  | sudo -u postgres psql
-echo "DROP USER inkspot_dns" | sudo -u postgres psql
+if [ -e /etc/inkspot/master_host ]; then
+	echo "Removing PowerDNS config and restarting..."
+	rm /etc/powerdns/pdns.d/inkspot.conf
+	service pdns restart
 
-echo "Removing system user..."
+	echo "Removing database and database users..."
+	echo "DROP DATABASE inkspot" | sudo -u postgres psql
+	echo "DROP USER inkspot"     | sudo -u postgres psql
+	echo "DROP USER inkspot_ro"  | sudo -u postgres psql
+	echo "DROP USER inkspot_dns" | sudo -u postgres psql
+
+fi
+
+echo "Removing inkspot system user..."
 userdel inkspot
 groupdel inkspot
 
 echo "Removing inKSpot installed files..."
 rm -rf /home/inkspot
 
-echo "Removing inkSpot user and comain homes..."
+echo "Removing inkSpot user and domain homes..."
 for file in `find /home/users -name .immutable`; do
 	chattr -i $file
 done
